Set custom interval to written metrics
Without this patch, threshold definitions for this plugin could raise
"NO DATA" notifications if an interval higher than the default one is
used in the plugin. This is caused by Collectd flushing the cache if
the stored metrics are older than the expected interval.

This patch sets the explicit interval set in the plugin when dispatching
the metrics. This makes Collectd aware of the custom Interval and not
raise a false notification due to no data being written.